project_aether/composer_dags/requirements.txt
astronomer-cosmos==1.7.1
apache-airflow-providers-snowflake==6.0.0
dbt-snowflake==1.10.0